Deleted projects and contexts now properly disappear in the Ajax 'fade' style when the delete button is clicked.
References to context.reset and project.reset are replaced by the correct Form.reset. Also replaced all instances of Form.focus_first with the correct Form.focusFirstElement.

Tidied up the edit forms on the projects and contexts pages. There's a weird visual bug in Safari (doesn't happen in Firefox), where the selected input box appears to be empty and unselected. However, if you type, your input will replace the previous text, and if you tab immediately to another field, you'll see that the input field was '''not''' actually empty.


git-svn-id: http://www.rousette.org.uk/svn/tracks-repos/trunk/tracks@150 a4c988fc-2ded-0310-b66e-134b36920a42