Fixes #243.
The 'done' page was generating errors (a regression in the trunk since the tagging of 1.04). I think that the main problem was the code to find the done items, but while I was at it I simplified the completed() and completed_archive() methods by using the Rails Time::Calculation methods x.day.ago and x.week.ago.

Works OK for me now.



git-svn-id: http://www.rousette.org.uk/svn/tracks-repos/trunk/tracks@213 a4c988fc-2ded-0310-b66e-134b36920a42