Added Luke's patch to fix the expand/collapse icons, which were formerly not working after an Ajax action had been performed without a refresh. This works properly now. Fixes #230.
I also modified the context pages slightly so that it only shows the last n completed actions in that project, where n is the number of completed actions user preference (no_completed). I'll do the same for projects. It prevents the context and project individual pages getting unmanageably long when you've been using it for a while.



git-svn-id: http://www.rousette.org.uk/svn/tracks-repos/trunk/tracks@214 a4c988fc-2ded-0310-b66e-134b36920a42