Fixes #403 "When prefs are set to date formats other than %m-%d-%y, database errors are produced on save."
This bug was due to the Chronic library integration, which I commented out for now.

In its youthful state, Chronic has a U.S.-centered worldview with only limited support for non-U.S. date formats. In particular, its tests show that it can support parsing 5/27/1979 and 27/5/1979 but are silent on the more challenging case of 12/11/2006 or 11/12/2006. I see no way to configure the plugin with date format. Without that, all users would have to use the (admittedly awkward) U.S. date format.

Maybe we could prefer a successful format-respecting date parse and fall back to a Chronic parse?

I also updated the "action due"" text on deferred items to use the user's date format.



git-svn-id: http://www.rousette.org.uk/svn/tracks-repos/trunk/tracks@358 a4c988fc-2ded-0310-b66e-134b36920a42